+++ cfgov/v1/models/sublanding_page.py
@@ -1,0 +1,53 @@
+from wagtail.wagtailcore.fields import StreamField
+from wagtail.wagtailadmin.edit_handlers import TabbedInterface, ObjectList, StreamFieldPanel
+from wagtail.wagtailcore import blocks
+from wagtail.wagtailimages.blocks import ImageChooserBlock
+
+from .base import CFGOVPage
+from . import molecules
+from . import organisms
+
+
+class SublandingPage(CFGOVPage):
+    content = StreamField([
+        ('header', blocks.CharBlock(icon='title')),
+        ('text_introduction', molecules.TextIntroduction()),
+        ('well', organisms.Well()),
+        ('hero', molecules.Hero()),
+        ('image_text_25_75', molecules.ImageText2575()),
+        ('image_text_50_50', molecules.ImageText5050()),
+        ('half_width_link_blob', molecules.HalfWidthLinkBlob()),
+        ('table', organisms.Table()),
+        ('contact', organisms.MainContactInfo()),
+    ], blank=True)
+    sidebar_breakout = StreamField([
+        ('slug', blocks.CharBlock(icon='title')),
+        ('heading', blocks.CharBlock(icon='title')),
+        ('paragraph', blocks.RichTextBlock(icon='edit')),
+        ('breakout_image', blocks.StructBlock([
+            ('image', ImageChooserBlock()),
+            ('is_round', blocks.BooleanBlock(required=False, default=True, label='Round?')),
+            ('icon', blocks.CharBlock(help_text='Enter icon class name.')),
+            ('heading', blocks.CharBlock(label='Introduction Heading')),
+            ('body', blocks.TextBlock(label='Introduction Body')),
+        ], heading='Breakout Image', icon='image')),
+        ('related_posts', organisms.RelatedPosts()),
+    ], blank=True)
+
+    # General content tab
+    content_panels = CFGOVPage.content_panels + [
+        StreamFieldPanel('content'),
+    ]
+
+    sidebar_panels = [
+        StreamFieldPanel('sidebar_breakout'),
+    ] + CFGOVPage.sidefoot_panels
+
+    # Tab handler interface
+    edit_handler = TabbedInterface([
+        ObjectList(content_panels, heading='General Content'),
+        ObjectList(sidebar_panels, heading='Sidebar'),
+        ObjectList(CFGOVPage.settings_panels, heading='Configuration'),
+    ])
+
+    template = 'sublanding-page/index.html'